refactor: renaming files and moving to utils
