--- src/main/java/aicore/pages/AbstractChangeAccessPopUpPage.java
@@ -1,37 +1,0 @@
-package aicore.pages;
-
-import com.microsoft.playwright.Page;
-
-import aicore.utils.ChangeAccessPopUpPageUtils;
-
-public abstract class AbstractChangeAccessPopUpPage {
-	protected Page page;
-
-	public String getChangeAccessPopupTitle() {
-		return ChangeAccessPopUpPageUtils.isChangeAccessPopupVisible(page);
-	}
-
-	public boolean isPopupVisible() {
-		return ChangeAccessPopUpPageUtils.isPopupVisible(page);
-	}
-
-	public boolean isOptionVisible(String option) {
-		return ChangeAccessPopUpPageUtils.isOptionVisible(page, option);
-	}
-
-	public void selectAccessType(String accessType) {
-		ChangeAccessPopUpPageUtils.selectAccessType(page, accessType);
-	}
-
-	public void enterComment(String comment) {
-		ChangeAccessPopUpPageUtils.enterComment(page, comment);
-	}
-
-	public void clickOnRequestButton() {
-		ChangeAccessPopUpPageUtils.clickOnRequestButton(page);
-	}
-
-	public String isRequestSuccessToastVisible() {
-		return ChangeAccessPopUpPageUtils.isRequestSuccessToastVisible(page);
-	}
-}
